Split SurfacePolice from UndergroundPolice
