+++ OPHD/Things/Structures/Structures.h
@@ -17,7 +17,8 @@
 #include "MineShaft.h"
 #include "Nursery.h"
 #include "Park.h"
-#include "Police.h"
+#include "SurfacePolice.h"
+#include "UndergroundPolice.h"
 #include "RecreationCenter.h"
 #include "RedLightDistrict.h"
 #include "Residence.h"
